refactor(podcast): 将类型定义提取到单独文件并优化展开逻辑
将Podcast相关类型定义提取到独立的type.ts文件
使用Set来管理展开状态以支持多条目独立展开
移除未使用的limit状态并修复分页计算
