+++ src/services/podcast/index.ts
@@ -1,36 +1,5 @@
 import request, { ErrorHandler } from '../request';
-
-// 获取博客列表
-export interface Podcast {
-  id: string;
-  title: string;
-  duration: number;
-  file_size: number;
-  audio_url: string;
-  content: string;
-  user: {
-    id: string;
-    name: string;
-    avatar_url: string;
-  };
-  created_at: string;
-  updated_at: string;
-}
-
-export interface ListParams {
-  page?: number;
-  limit?: number;
-}
-
-export interface ListResponse<T> {
-  podcasts: T[];
-  total: number;
-  totalPages: number;
-  hasNext: boolean;
-  hasPrev: boolean;
-  limit: number;
-  page: number;
-}
+import { Podcast, ListParams, ListResponse } from './type';
 
 export const PodcastApi = {
   getList: (params: ListParams = { page: 1, limit: 10 }, errorHandler?: ErrorHandler) =>
@@ -48,7 +17,6 @@ export const PodcastApi = {
       errorHandler,
       timeout: 150000,
       params: formData,
-      // headers: { 'Content-Type': 'multipart/form-data' },
     }),
 };
